+++ src/main/kotlin/site/siredvin/peripheralium/util/representation/LuaRepresentation.kt
@@ -8,18 +8,21 @@ import net.minecraft.world.effect.MobEffect
 import net.minecraft.world.effect.MobEffectInstance
 import net.minecraft.world.entity.Entity
 import net.minecraft.world.entity.npc.Villager
-import net.minecraft.world.entity.player.Player
+import net.minecraft.world.entity.npc.VillagerProfession
+import net.minecraft.world.entity.npc.WanderingTrader
 import net.minecraft.world.item.Item
 import net.minecraft.world.item.ItemStack
 import net.minecraft.world.item.Items
 import net.minecraft.world.item.enchantment.Enchantment
 import net.minecraft.world.item.enchantment.EnchantmentHelper
+import net.minecraft.world.item.trading.Merchant
 import net.minecraft.world.item.trading.MerchantOffer
 import net.minecraft.world.level.block.state.BlockState
 import site.siredvin.peripheralium.ext.toRelative
 import java.util.stream.Collectors
 import java.util.stream.Stream
 
+@Suppress("MemberVisibilityCanBePrivate")
 object LuaRepresentation {
 
     fun forBlockState(state: BlockState): MutableMap<String, Any> {
@@ -61,11 +64,32 @@ object LuaRepresentation {
         return map
     }
 
+    fun forEnchantment(enchantment: Enchantment, level: Int = 1): MutableMap<String, Any> {
+        return mutableMapOf(
+            "name" to enchantment.getFullname(level).string,
+            "technicalName" to fromLegacyToNewID(enchantment.descriptionId),
+            "level" to level
+        )
+    }
+
+    fun forEnchantments(enchantments : MutableMap<Enchantment, Int>): List<Map<String, Any>> {
+        val list = mutableListOf<Map<String, Any>>()
+        for(enchantment:MutableMap.MutableEntry<Enchantment, Int> in enchantments.entries) {
+            list.add(forEnchantment(enchantment.key, enchantment.value))
+        }
+        return list
+    }
+
     fun forItemStack(stack: ItemStack): MutableMap<String, Any> {
         val map = forItem(stack.item)
         map["tags"] = tagsToList(stack.tags)
         map["count"] = stack.count
         map["maxStackSize"] = stack.maxStackSize
+        if(stack.isEnchanted || stack.`is`(Items.ENCHANTED_BOOK)) {
+            val enchantments = EnchantmentHelper.getEnchantments(stack)
+            if (enchantments != null)
+                map["enchantments"] = forEnchantments(enchantments)
+        }
         return map
     }
 
@@ -79,7 +103,7 @@ object LuaRepresentation {
     fun forMobEffect(effect: MobEffect): MutableMap<String, Any> {
         return hashMapOf(
             "name" to effect.displayName.string,
-            "technicalName" to effect.descriptionId,
+            "technicalName" to fromLegacyToNewID(effect.descriptionId),
         )
     }
 
@@ -97,38 +121,44 @@ object LuaRepresentation {
         return tags.map { key -> key.location.toString() }.collect(Collectors.toList())
     }
 
-    fun getVillagerOffersAsMap(villager: Villager): MutableMap<Int, Any> {
-        val offersMap : MutableMap<Int, Any> = HashMap()
-        var index: Int = 1
-        for(merchantOffer:MerchantOffer in villager.offers) {
+    fun forMerchantOffers(merchant: Merchant): Map<Int, Map<String, Any>> {
+        val offers = mutableMapOf<Int, Map<String, Any>>()
+        var currentIndex = 1
+        for(merchantOffer:MerchantOffer in merchant.offers) {
+            if (merchantOffer.isOutOfStock) {
+                currentIndex++
+                continue
+            }
             val offerMap : MutableMap<String, Any> = HashMap()
-            offerMap["first_cost"] = getItemStackAsMap(merchantOffer.costA)
-            offerMap["second_cost"] = getItemStackAsMap(merchantOffer.costB)
-            offerMap["result"] = getItemStackAsMap(merchantOffer.result)
-            offersMap[index++] = offerMap
+            val inputs: MutableList<Map<String, Any>> = mutableListOf()
+            inputs.add(forItemStack(merchantOffer.costA))
+            if (!merchantOffer.costB.isEmpty)
+                inputs.add(forItemStack(merchantOffer.costB))
+            offerMap["inputs"] = inputs
+            offerMap["outputs"] = listOf(forItemStack(merchantOffer.result))
+            offers[currentIndex] = offerMap
+            currentIndex++
         }
-        return offersMap
-    }
-    private fun getItemStackAsMap(itemStack: ItemStack): MutableMap<String, Any> {
-        val itemStackMap : MutableMap<String, Any> = HashMap()
-        itemStackMap["count"] = itemStack.count
-//        var name = itemStack.displayName.string
-//        name = name.substring(0, name.length-1).substring(1, )
-        itemStackMap["name"] = formatID(itemStack.descriptionId)
-        if(itemStack.isEnchanted || itemStack.`is`(Items.ENCHANTED_BOOK)) itemStackMap["enchantments"] = getEnchantmentsAsMap(itemStack)
-        return itemStackMap
+        return offers
     }
 
-    private fun getEnchantmentsAsMap(itemStack : ItemStack): MutableMap<String, Any> {
-        val enchantmentsMap : MutableMap<String, Any> = HashMap()
-        val enchantments: MutableMap<Enchantment, Int>? = EnchantmentHelper.getEnchantments(itemStack)
-        for(enchantment:MutableMap.MutableEntry<Enchantment, Int> in enchantments!!.entries) {
-            enchantmentsMap[formatID(enchantment.key.descriptionId)] = enchantment.value
+    fun forVillager(villager: Villager): Map<String, Any> {
+        val data = mutableMapOf<String, Any>()
+        val vilData = villager.villagerData
+        if (vilData.profession != VillagerProfession.NONE) {
+            data["profession"] = vilData.profession.name
+            data["xp"] = villager.villagerXp
+            data["level"] = vilData.level
+            data["type"] = vilData.type.toString()
         }
-        return enchantmentsMap
+        return data
     }
 
-    private fun formatID(rawID: String): String {
-        return rawID.substring(rawID.indexOf(".") + 1).replace(".", ":")
+    /**
+     * So, this function exists mostly for converting ids like minecraft.looting to more
+     * simple for anyone minecraft:looting. Mostly applicable for enchantments and effects
+     */
+    private fun fromLegacyToNewID(legacyID: String): String {
+        return legacyID.substring(legacyID.indexOf(".") + 1).replace(".", ":")
     }
 }